+++ providers/offline/src/index.ts
@@ -1,3 +1,5 @@
+export * from './providers/deck.offline.provider';
+export * from './providers/slide.offline.provider';
 export * from './providers/doc.offline.provider';
 export * from './providers/paragraph.offline.provider';
 export * from './providers/editor.provider';