+++ providers/offline/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.0.4 (2022-02-25)
+
+### Features
+
+- offline slide providers
+- 
+# 0.0.3 (2022-02-25)
+
+### Features
+
+- offline deck providers
+
 # 0.0.2 (2022-02-24)
 
 ### Build